Add livestage field and React renderer
