+++ liveblog/blocks.py
@@ -1,0 +1,4 @@
+from wagtail.blocks import StructBlock
+
+class LiveblogHeader(StructBlock):
+    pass
